RDKSYSINTSW-3547 : Set/verify/analyze/ single file metadata on a target
Reason for change: Provide support to rootFSTTSAnalyzer.sh script to
set/verify/analyze/ metadata of a rootFS single file or sub-folder on a
target platform.
Test Procedure: Execute rootFSTTSAnalyzer.sh script on a target platform
with a mutualy exclusive option {-s|-v|-a} to set/verify/analyze a single
file, sub-folder, or an entire rootFS folder metadata. In first 2 cases,
a path to a version file must be explicitely set.
Risks: None.
Signed-off-by: Sergei Danilchuk <sergei_danilchuk@cable.comcast.com>

Change-Id: Ib0d23ad83ba0695b71200ce4d6a67955dd1729d7